Safari-609.1.11.1/LayoutTests/fast/dom/navigation-with-sideeffects-expected.txt

ALERT: completed o1.toString()
ALERT: completed o2.toString
